rename Version.h to StaticData.h new temp file path for windows and linux temp file now writeable on windows compare string find method with SIZE_MAX instead of ULONG_MAX

+++ src/IpHelper.cpp
@@ -1,7 +1,5 @@
 #include "IpHelper.h"
 
-#include <climits>
-
 bool IpHelper::isIpValid(std::string ip) {
-    return (ip.find('.') != ULONG_MAX);
+    return (ip.find('.') != SIZE_MAX);
 }
